repository 생성 및 프로젝트 구조 변경
