+++ src/main/java/capstone/recipable/domain/bookmark/repository/BookmarkRepository.java
@@ -1,0 +1,7 @@
+package capstone.recipable.domain.bookmark.repository;
+
+import capstone.recipable.domain.bookmark.entity.Bookmark;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
+}
