+++ src/main/java/capstone/recipable/domain/category/entity/Category.java
@@ -1,26 +1,27 @@
-package capstone.recipable.domain.expiration;
+package capstone.recipable.domain.category.entity;
 
-import capstone.recipable.domain.ingredient.Ingredient;
+import capstone.recipable.domain.refrigerator.entity.Refrigerator;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDate;
-
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class Expiration {
+public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private LocalDate expireDate;
+    private String categoryName;
 
-    @OneToOne
-    private Ingredient ingredientId;
+    private String details;
+
+    @ManyToOne(fetch=FetchType.LAZY)
+    private Refrigerator refrigeratorId;
+
 }
