+++ src/main/java/capstone/recipable/domain/category/repository/CategoryRepository.java
@@ -1,0 +1,7 @@
+package capstone.recipable.domain.category.repository;
+
+import capstone.recipable.domain.category.entity.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+}
