+++ src/main/java/capstone/recipable/domain/expiration/repository/ExpirationRepository.java
@@ -1,0 +1,7 @@
+package capstone.recipable.domain.expiration.repository;
+
+import capstone.recipable.domain.expiration.entity.Expiration;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ExpirationRepository extends JpaRepository<Expiration, Long> {
+}
