+++ src/main/java/capstone/recipable/domain/ingredient/repository/IngredientRepository.java
@@ -1,0 +1,7 @@
+package capstone.recipable.domain.ingredient.repository;
+
+import capstone.recipable.domain.ingredient.entity.Ingredient;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
+}
