+++ src/main/java/capstone/recipable/domain/recipe/repository/RecipeRepository.java
@@ -1,0 +1,7 @@
+package capstone.recipable.domain.recipe.repository;
+
+import capstone.recipable.domain.recipe.entity.Recipe;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RecipeRepository extends JpaRepository<Recipe, Long> {
+}
