+++ src/main/java/capstone/recipable/domain/refrigerator/repository/RefrigeratorRepository.java
@@ -1,0 +1,7 @@
+package capstone.recipable.domain.refrigerator.repository;
+
+import capstone.recipable.domain.refrigerator.entity.Refrigerator;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RefrigeratorRepository extends JpaRepository<Refrigerator, Long> {
+}
