+++ src/main/java/capstone/recipable/domain/user/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package capstone.recipable.domain.user.repository;
+
+import capstone.recipable.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
